B!! All the version interactions are working in SQLite now. Fixed a syntax error in the version check script, changed version storage to longs, made other changes along the way. Now to go back to another branch and see if I can do this without doing multiple DB hits each time I get the version.

+++ DbMigrate.Tests/CommonInfrastructureForAllMmfs/DatabaseManageDbVersion.cs
@@ -10,8 +10,8 @@ namespace DbMigrate.Tests.CommonInfrastructureForAllMmfs
 	{
 		private const string CreateVersionInfoTableSql =
             @"create table __database_info(
-  min_version_number int,
-  max_version_number int
+  min_version_number long,
+  max_version_number long
 );
 insert into __database_info(min_version_number, max_version_number) values(0, 0);";
